+++ tests/unit/cli-env-inline-comment-10100.test.ts
@@ -2,42 +2,13 @@ import test from "node:test";
 import assert from "node:assert/strict";
 import fs from "node:fs";
 import path from "node:path";
+import { parseEnvValue } from "../../bin/cli/utils/parseEnvValue.mjs";
 
 // #10100 — the .env loader kept inline comments inside values, so the shipped
 // `QUOTA_STORE_DRIVER=sqlite  # sqlite | redis` line produced the literal value
 // "sqlite              # sqlite | redis". Consumers compare with `===`, so a
 // user annotating `QUOTA_STORE_DRIVER=redis  # ...` silently got SQLite with no
 // warning (the existing warning lives inside the `redis` branch).
-
-const LOADER = path.resolve("bin/omniroute.mjs");
-
-/**
- * The loader is a CLI entrypoint with side effects on import, so exercise the
- * pure helper by extracting it from source rather than importing the module.
- */
-function loadParseEnvValue(): (raw: string) => string {
-  const source = fs.readFileSync(LOADER, "utf8");
-  const start = source.indexOf("function parseEnvValue(");
-  assert.ok(start > -1, "parseEnvValue should exist in bin/omniroute.mjs");
-  // Walk to the end of the function body.
-  let depth = 0;
-  let end = start;
-  for (let i = source.indexOf("{", start); i < source.length; i++) {
-    if (source[i] === "{") depth++;
-    else if (source[i] === "}") {
-      depth--;
-      if (depth === 0) {
-        end = i + 1;
-        break;
-      }
-    }
-  }
-  return new Function(`${source.slice(start, end)}; return parseEnvValue;`)() as (
-    raw: string
-  ) => string;
-}
-
-const parseEnvValue = loadParseEnvValue();
 
 test("an unquoted inline comment is stripped", () => {
   assert.equal(parseEnvValue("sqlite              # sqlite | redis"), "sqlite");
